% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmac2.R
\name{bench}
\alias{bench}
\title{Calculate patient volume stay benchmarks for SurgeSim}
\usage{
bench(db = "medstat1_surgesim", protocol)
}
\arguments{
\item{db}{Simulation database containing datamine}

\item{protocol}{Simulation protocol for benchmarks}
}
\description{
Calculate patient volume stay benchmarks for SurgeSim
}
\examples{
z<-bench(protocol='Geyserville');
}

