% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmac2.R
\name{pv}
\alias{pv}
\title{Calculate patient volume stay benchmarks for SurgeSim}
\usage{
pv(db = "medstat1_surgesim", sim, protocol)
}
\arguments{
\item{db}{Main Simulation database containing datamine}

\item{sim}{Simulation database}

\item{protocol}{Simulation protocol for comparison}
}
\description{
Calculate patient volume stay benchmarks for SurgeSim
}
\examples{
z<-pv(protocol='Geyserville',sim='unmert');
}

