% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage.R
\name{triage_accuracy}
\alias{triage_accuracy}
\title{Calculate Triage Accuracy for SurgeSim}
\usage{
triage_accuracy(sim, model)
}
\arguments{
\item{sim}{Simulation database}

\item{model}{Triage method used: 'start','ctas', or 'nato'}
}
\description{
Calculate Triage Accuracy for SurgeSim
}
\examples{
z<-triage_accuracy(sim='emdm2016a',model='start');
}

