% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmac2.R
\name{los}
\alias{los}
\title{Calculate length of stay benchmarks for SurgeSim}
\usage{
los(db = "medstat1_surgesim", sim, protocol)
}
\arguments{
\item{db}{Main Simulation database containing datamine}

\item{sim}{Simulation database}

\item{protocol}{Simulation protocol for comparison}
}
\description{
Calculate length of stay benchmarks for SurgeSim
}
\examples{
z<-los(protocol='Geyserville',sim='unmert');
}

