% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydf2.R
\name{mydf2config_make}
\alias{mydf2config_make}
\title{Write configuration object for the mydf2 function.}
\usage{
mydf2config_make(user, password, host, db = NULL)
}
\arguments{
\item{user}{Default username.}

\item{password}{Password for default user.}

\item{host}{Hostname for MySQL database}

\item{db}{Optional default database}
}
\value{
Null.  Used for its side effect of writing the mydf2config object
}
\description{
Write configuration object for the mydf2 function.
}
\examples{
mydf2config_make(user='jeffrey',password='34tgh',host='localhost',db='foo')
mydf2config_make(user='jeffrey',password='34tgh',host='localhost')
}

