% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radmac2.R
\docType{package}
\name{radmac2}
\alias{radmac2}
\alias{radmac2-package}
\title{radmac2: Calculate and plot surge capacity metrics from SurgeSim}
\description{
The package provides four types of benchmark plots
}
\details{
The package requires a configuration object mydf2config which can be written
to the environment with the function mydf2config_make
}
\references{
Jeffrey Michael Franc, Pier Luigi Ingrassia, Manuela Verde, Davide Colombo
and Francesco Della Corte (2015). A Simple Graphical Method for Quantification
of Disaster Management Surge Capacity Using Computer Simulation and Process-
control Tools. Prehospital and Disaster Medicine, 30, pp 9-15.
}

